Added and updated a lot since the last checkin; too much to mention.

+++ trunk/SnowGlobe/Snowglobe/Quad.h
@@ -14,22 +14,22 @@
 class Quad : public IGeometry
 {
 private:
-	bool			m_bInitialized;
-	CustomVertex *	m_pVertices;
-	DWORD *			m_pdwIndices;
+	bool				m_bInitialized;
+	CustomVertex *		m_pVertices;
+	unsigned short *	m_pdwIndices;
 
-	float			m_rWidth;
-	float			m_rHeight;
+	float				m_rWidth;
+	float				m_rHeight;
 
-	int				m_nVertCount;
-	int				m_nIndexCount;
+	int					m_nVertCount;
+	int					m_nIndexCount;
 
 private:
 	bool Initialize();
 	
 	// IGeometry
 	HRESULT GenerateVertices( CustomVertex** ppVertices );
-	HRESULT GenerateIndices( DWORD** ppIndices );
+	HRESULT GenerateIndices( unsigned short** ppIndices );
 
 public:
 	Quad();
@@ -47,5 +47,5 @@ public:
 	const int TriCount() const		{ return m_nIndexCount / 3; }
 
 	CustomVertex** Vertices()		{ return &m_pVertices; }
-	DWORD* Indices()				{ return m_pdwIndices; }
+	unsigned short* Indices()		{ return m_pdwIndices; }
 };
